fix: correct RealFunction type declaration

+++ include/types.h
@@ -2,4 +2,4 @@
  * Type definitions for mathematical entities
  */
 
-typedef double (*RealFunction)(double *x, float d);
+typedef double (*RealFunction)(double *x, unsigned int d);
